pyproject.toml
[tool.poetry]
name = "tea-xai-demo"
version = "0.1.0"
description = ""
authors = ["Christopher Burr <chrisdburr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pandas = "^2.2.3"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
markdown = "^3.7"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
html5lib = "^1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-extensions = "^3.2.3"
django-tailwind = "^3.8.0"
django-browser-reload = "^1.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
